Kontrola pristupa i biranje uloga korisnika
Dodana mogućnost mijenjanja uloge CLIENT/SELLER/ADMIN. Dodana provjera uloge i valjanosti pretplate kod postavljanja oglasa.

+++ src/gearshare/src/main/java/com/gearshare/gearshare/domain/entities/SuspensionEntity.java
@@ -33,6 +33,6 @@ public class SuspensionEntity {
     private LocalDateTime suspensionStartDateTime;
 
     @NotNull
-    private int suspensionLength;
+    private short suspensionLength;
 
 }
